requirements.txt
flask==2.0.1
werkzeug==2.0.3
torch==2.0.1
torchvision==0.15.2
transformers==4.36.2
Pillow==9.5.0
numpy==1.24.3
timm==0.9.2
einops==0.6.1
modelscope==1.9.5
qwen==1.0.0 